requirements.txt

paho-mqtt ~= 1.6.0
sqlalchemy > =1.4.44
mysqlclient ~= 2.1.0
python-dotenv ~= 0.21.0